requirements.txt
defusedxml==0.6.0
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-xml==1.4.0
pytz==2019.3
sqlparse==0.3.0
